Refactor: Remove Bottom Sheet components and update dependencies
This commit removes the following Bottom Sheet components from the `presentation` module:
- `MovieListsBottomSheet.kt`
- `RatingMovieBottomSheet.kt`
- `SaveMovieBottomSheet.kt`
- `ShareMovieBottomSheet.kt`

**Dependency Updates:**
- Removed `androidx.material3` and `androidx.material3.android` dependencies from `presentation/build.gradle.kts` as they are no longer directly used after the bottom sheet removals.

**Component Relocation:**
- Moved `ExpandableDescription.kt` from `presentation/screens/detailsScreen/componant` to `designSystem/component`.

**UI Updates:**
- Replaced `Text` with `MovioText` in `ForYouAndExploreScreen.kt` and `TrendingLayout.kt` for consistency.
- Updated `MovioButton.kt` in `designSystem` to:
    - Add an optional `backgroundColor` parameter. If `null`, the button will have no background and minimal padding, suitable for text buttons.
    - Introduce a new `MovioTextButton` composable for creating text-only buttons.

**Impact:**
- The removed Bottom Sheet components were specific to the `detailsScreen` and are no longer part of the application.
- The `ExpandableDescription` component is now part of the `designSystem` for broader reusability.
- UI components in search and home screens now use the consistent `MovioText` component.
- `MovioButton` offers more flexibility in styling, particularly for text buttons.
