+++ designSystem/src/main/java/com/madrid/designSystem/component/MovioButton.kt
@@ -6,6 +6,7 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -15,20 +16,28 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import com.madrid.designSystem.theme.Theme
 
-
 @Composable
 fun MovioButton(
     modifier: Modifier = Modifier,
     color: Color = Theme.color.brand.primary,
+    backgroundColor: Color? = color,
     onClick: () -> Unit,
     content: @Composable RowScope.() -> Unit,
 ) {
-    Box(
-
-        modifier = modifier
+    val buttonModifier = if (backgroundColor != null) {
+        modifier
             .clip(RoundedCornerShape(32.dp))
-            .background(color)
-            .clickable { onClick() },
+            .background(backgroundColor)
+            .clickable { onClick() }
+            .padding(horizontal = 16.dp, vertical = 8.dp)
+    } else {
+        modifier
+            .clickable { onClick() }
+            .padding(horizontal = 4.dp, vertical = 4.dp)
+    }
+
+    Box(
+        modifier = buttonModifier,
         contentAlignment = Alignment.Center
     ) {
         Row(
@@ -37,4 +46,19 @@ fun MovioButton(
             content = content
         )
     }
+}
+@Composable
+fun MovioTextButton(
+    text: String,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+    textColor: Color = Theme.color.brand.primary,
+    textStyle: androidx.compose.ui.text.TextStyle = Theme.textStyle.label.smallRegular14
+) {
+    MovioText(
+        text = text,
+        color = textColor,
+        textStyle = textStyle,
+        modifier = modifier.clickable { onClick() }
+    )
 }
